requirements.txt
google-cloud-firestore
google-generativeai
httpx
python-dotenv
moviepy
tiktok-uploader
sentry-sdk
